+++ src/libs/kubernetes_checker.py
@@ -15,9 +15,9 @@ class KubernetesChecker:
                  debug_on=True,
                  logger=None,
                  queue=None,
-                 telegram_queue=None,
-                 telegram_max_msg_len=2000,
-                 telegram_alive_message_hours=0,
+                 dispatcher_queue=None,
+                 dispatcher_max_msg_len=8000,
+                 dispatcher_alive_message_hours=24,
                  k8s_key_config: ConfigK8sProcess = None):
 
         self.print_helper = PrintHelper('k8s_checker', logger)
@@ -28,8 +28,8 @@ class KubernetesChecker:
 
         self.queue = queue
 
-        self.telegram_max_msg_len = telegram_max_msg_len
-        self.telegram_queue = telegram_queue
+        self.dispatcher_max_msg_len = dispatcher_max_msg_len
+        self.dispatcher_queue = dispatcher_queue
 
         self.k8s_config = ConfigK8sProcess()
         if k8s_key_config is not None:
@@ -44,11 +44,20 @@ class KubernetesChecker:
         self.old_pv = {}
         self.old_deployment = {}
 
-        self.alive_message_seconds = telegram_alive_message_hours * 3600
+        self.alive_message_seconds = dispatcher_alive_message_hours * 3600
         self.last_send = calendar.timegm(datetime.today().timetuple())
 
         self.cluster_name = ""
         self.force_alive_message = False
+
+        # LS 2023.11.02 add variable for sending isolate message configuration
+        self.send_config = False
+
+        # LS 2023.11.03 add final message for concat data all keys in a one message
+        # Not required a unique message
+        self.final_message = ""
+        # Not required a unique message
+        self.unique_message = False
 
     @handle_exceptions_async_method
     async def __put_in_queue__(self,
@@ -56,23 +65,50 @@ class KubernetesChecker:
                                obj):
         """
         Add new element to the queue
-        :param queue:
-        :param obj:
+        :param queue: reference to a queue
+        :param obj: objet to add
         """
         self.print_helper.info_if(self.print_debug, "__put_in_queue__")
 
         await queue.put(obj)
 
     @handle_exceptions_async_method
-    async def send_to_telegram(self, message):
+    async def send_to_dispatcher(self, message):
         """
-        Send message to Telegram engine
-        :param message:
+        Send message to dispatcher engine
+        @param message: message to send
         """
-        self.print_helper.info(f"send_to_telegram")
-        self.last_send = calendar.timegm(datetime.today().timetuple())
-        await self.__put_in_queue__(self.telegram_queue,
-                                    message)
+        self.print_helper.info(f"send_to_dispatcher. msg len= {len(message)}")
+        if len(message) > 0:
+            if not self.unique_message:
+                self.last_send = calendar.timegm(datetime.today().timetuple())
+                await self.__put_in_queue__(self.dispatcher_queue,
+                                            message)
+            else:
+
+                if len(self.final_message) > 0:
+                    self.print_helper.info(f"send_to_dispatcher. concat message- len({len(self.final_message)})")
+                    self.final_message = f"{self.final_message}\n{'-'*20}\n{message}"
+                else:
+                    self.print_helper.info(f"send_to_dispatcher. start message")
+                    self.final_message = f"{message}"
+
+    @handle_exceptions_async_method
+    async def send_to_dispatcher_summary(self):
+        """
+        Send summary message to dispatcher engine
+        """
+
+        self.print_helper.info(f"send_to_dispatcher_summary. message-len= {len(self.final_message)}")
+        # Chck if the final message is not empty
+        if len(self.final_message) > 10:
+            self.final_message = f"Start report\n{self.final_message}\nEnd report"
+            self.last_send = calendar.timegm(datetime.today().timetuple())
+            await self.__put_in_queue__(self.dispatcher_queue,
+                                        self.final_message)
+
+        self.final_message = ""
+        self.unique_message = False
 
     @handle_exceptions_async_method
     async def __process_key__(self,
@@ -139,10 +175,10 @@ class KubernetesChecker:
                         # if resolved:
                         #     msg += f"Resolved= True\n"
 
-                        if len(msg) >= self.telegram_max_msg_len:
+                        if len(msg) >= self.dispatcher_max_msg_len:
                             self.print_helper.info_if(self.print_debug,
                                                       f"Max message length reached, force send message")
-                            await self.send_to_telegram(msg)
+                            await self.send_to_dispatcher(msg)
                             # Reset msg len
                             msg = base_title
 
@@ -153,7 +189,7 @@ class KubernetesChecker:
                 if len(msg) > len(base_title):
                     self.print_helper.info_if(self.print_debug, f"Flush last message")
 
-                    await self.send_to_telegram(msg)
+                    await self.send_to_dispatcher(msg)
 
                 if (not resolved
                         and data is not None
@@ -172,7 +208,14 @@ class KubernetesChecker:
             self.print_helper.error_and_exception(f"__process_key__{title}", err)
 
     async def __concatenate__status__(self, key, key_value, msg, value):
-
+        """
+        concatenate the message. Iterate in the f
+        @param key: title of section
+        @param key_value:
+        @param msg: concatenate variable
+        @param value: dict value
+        @return:
+        """
         if isinstance(value, dict):
             key_value = False
             msg += f"{key}:\n"
@@ -215,23 +258,30 @@ class KubernetesChecker:
                     await self.__process_pv__(data)
                 elif self.k8s_config.PVC_key in data:
                     await self.__process_pvc__(data)
+                # LS 2023.11.03 add key message for sending unique message
+                elif self.k8s_config.disp_MSG_key_start in data:
+                    self.unique_message = True
+                    self.final_message = ""
+                # LS 2023.11.03 add key message for sending unique message
+                elif self.k8s_config.disp_MSG_key_end in data:
+                    await self.send_to_dispatcher_summary()
                 else:
                     self.print_helper.info(f"key not defined")
             else:
                 self.print_helper.info(f"__unpack_data.the message is not a type of dict")
 
-            # telegram alive message
+            # dispatcher alive message
             if self.alive_message_seconds > 0:
                 diff = calendar.timegm(datetime.today().timetuple()) - self.last_send
 
                 if diff > self.alive_message_seconds or self.force_alive_message:
                     self.print_helper.info(f"__unpack_data.send alive message")
-                    await self.send_to_telegram(f"Cluster: {self.cluster_name}"
-                                                f"\nk8s-watchdog is running."
-                                                f"\nThis is an alive message"
-                                                f"\nNo warning/errors were triggered in the last "
-                                                f"{int(self.alive_message_seconds/3600)} "
-                                                f"hours ")
+                    await self.send_to_dispatcher(f"Cluster: {self.cluster_name}"
+                                                  f"\nk8s-watchdog is running."
+                                                  f"\nThis is an alive message"
+                                                  f"\nNo warning/errors were triggered in the last "
+                                                  f"{int(self.alive_message_seconds / 3600)} "
+                                                  f"hours ")
                     self.force_alive_message = False
 
         except Exception as err:
@@ -260,6 +310,10 @@ class KubernetesChecker:
 
     @handle_exceptions_async_method
     async def __process_cluster_name__(self, data):
+        """
+        Obtain cluster name
+        @param data:
+        """
         self.print_helper.info(f"__process_cluster_name__")
 
         nodes_name = data[self.k8s_config.CLUSTER_Name_key]
@@ -267,7 +321,11 @@ class KubernetesChecker:
         self.print_helper.info(f"cluster name {nodes_name}")
         if nodes_name is not None:
             self.print_helper.info_if(self.print_debug, f"Flush last message")
-            await self.send_to_telegram(f"Cluster name= {nodes_name}")
+            # LS 2023.11.04 Send configuration separately
+            if self.send_config:
+                await self.send_to_dispatcher(f"Cluster name= {nodes_name}")
+            else:
+                await self.send_active_configuration(f"Cluster name= {nodes_name}")
 
         self.cluster_name = nodes_name
 
@@ -430,11 +488,16 @@ class KubernetesChecker:
         self.old_pvc = pvc_status
 
     @handle_exceptions_async_method
-    async def send_active_configuration(self):
+    async def send_active_configuration(self, sub_title=None):
         """
         Send a message to Telegram engine of the active setup
         """
+        title = "k8s-watchdog is restarted"
+        if sub_title is not None and len(sub_title) > 0:
+            title = f"{title}\n{sub_title}"
+
         self.print_helper.info_if(self.print_debug, f"send_active_configuration")
+
         msg = f'Configuration setup:\n'
         if self.k8s_config is not None:
             msg = msg + f"  . node status= {'ENABLE' if self.k8s_config.NODE_enable else '.'}\n"
@@ -442,21 +505,22 @@ class KubernetesChecker:
             msg = msg + (f"  . deployment= {'ENABLE' if self.k8s_config.DPL_enable else '.'} "
                          f"{'P0' if self.k8s_config.DPL_enable and self.k8s_config.DPL_pods0 else ''}\n")
             msg = msg + (f"  . stateful sets= {'ENABLE' if self.k8s_config.SS_enable else '.'} "
-                         f"{'P0' if self.k8s_config.SS_enable  and self.k8s_config.SS_pods0 else ''}\n")
-            msg = msg + (f"  . replicaset= {'ENABLE' if  self.k8s_config.RS_enable else '.'} "
+                         f"{'P0' if self.k8s_config.SS_enable and self.k8s_config.SS_pods0 else ''}\n")
+            msg = msg + (f"  . replicaset= {'ENABLE' if self.k8s_config.RS_enable else '.'} "
                          f"{'P0' if self.k8s_config.RS_enable and self.k8s_config.RS_pods0 else ''}\n")
             msg = msg + (f"  . daemon sets= {'ENABLE' if self.k8s_config.DS_enable else '-'}"
-                         f"{'P0' if  self.k8s_config.DS_enable and  self.k8s_config.DS_pods0 else ''}\n")
+                         f"{'P0' if self.k8s_config.DS_enable and self.k8s_config.DS_pods0 else ''}\n")
             msg = msg + f"  . pvc= {'ENABLE' if self.k8s_config.PVC_enable else '.'}\n"
             msg = msg + f"  . pv= {'ENABLE' if self.k8s_config.PVC_enable else '.'}\n"
             if self.alive_message_seconds >= 3600:
-                msg = msg + f"\nAlive message every {int(self.alive_message_seconds/3600)} hours"
+                msg = msg + f"\nAlive message every {int(self.alive_message_seconds / 3600)} hours"
             else:
                 msg = msg + f"\nAlive message every {int(self.alive_message_seconds / 60)} minutes"
         else:
             msg = "Error init config class"
 
-        await self.send_to_telegram(f"k8s-watchdog is restarted\n\n{msg}")
+        msg = f"{title}\n\n{msg}"
+        await self.send_to_dispatcher(msg)
 
     @handle_exceptions_async_method
     async def run(self):
@@ -465,7 +529,8 @@ class KubernetesChecker:
         """
         try:
             self.print_helper.info("checker run")
-            await self.send_active_configuration()
+            if self.send_config:
+                await self.send_active_configuration()
 
             while True:
                 # get a unit of work