+++ src/utils/config.py
@@ -98,7 +98,7 @@ class ConfigProgram:
     @handle_exceptions_method
     def telegram_max_msg_len(self):
         res = self.load_key('TELEGRAM_MAX_MSG_LEN',
-                            '2000')
+                            '3000')
 
         if len(res) == 0:
             res = '2000'
@@ -114,8 +114,8 @@ class ConfigProgram:
         return int(res)
 
     @handle_exceptions_method
-    def telegram_alive_message_hours(self):
-        res = self.load_key('TELEGRAM_ALIVE_MSG_HOURS',
+    def notification_alive_message_hours(self):
+        res = self.load_key('NOTIFICATION_ALIVE_MSG_HOURS',
                             '24')
         n_hours = int(res)
         if n_hours < 0:
@@ -124,6 +124,34 @@ class ConfigProgram:
             n_hours = 100
 
         return n_hours
+
+    @handle_exceptions_method
+    def email_enable(self):
+        res = self.load_key('EMAIL_ENABLE', 'False')
+        return True if res.lower() == "true" or res.lower() == "1" else False
+
+    @handle_exceptions_method
+    def email_smtp_server(self):
+        return self.load_key('EMAIL_SMTP_SERVER', '')
+
+    @handle_exceptions_method
+    def email_smtp_port(self):
+        res = self.load_key('EMAIL_SMTP_PORT',
+                            '587')
+        n_port = int(res)
+        return n_port
+
+    @handle_exceptions_method
+    def email_recipient(self):
+        return self.load_key('EMAIL_RECIPIENTS', '')
+
+    @handle_exceptions_method
+    def email_sender_password(self):
+        return self.load_key('EMAIL_PASSWORD', '', mask_value=True)
+
+    @handle_exceptions_method
+    def email_sender(self):
+        return self.load_key('EMAIL_ACCOUNT', '')
 
     @handle_exceptions_method
     def k8s_load_kube_config_method(self):
@@ -231,6 +259,12 @@ class ConfigK8sProcess:
 
         self.NODE_enable = True
         self.NODE_key = 'nodelist'
+
+        # LS 2023.11.03 key for sending a unique message
+        self.disp_MSG_key_unique = True  # Fixed True
+        self.disp_MSG_key_start = 'msg_key_start'
+        self.disp_MSG_key_end = 'msg_key_end'
+
         if cl_config is not None:
             self.__init_configuration_app__(cl_config)
 
@@ -250,6 +284,8 @@ class ConfigK8sProcess:
         print(f"INFO    [Process setup] k8s check pvc={self.PVC_enable}")
         print(f"INFO    [Process setup] k8s check pv={self.PVC_enable}")
 
+        print(f"INFO    [Process setup] k8s send summary message={self.disp_MSG_key_unique}")
+
     def __init_configuration_app__(self, cl_config: ConfigProgram):
         """
         Init configuration class reading .env file
@@ -268,4 +304,87 @@ class ConfigK8sProcess:
         self.RS_pods0 = cl_config.k8s_replica_sets_pods0()
         self.CLUSTER_Name_forced = cl_config.k8s_force_cluster_identification()
 
+        self.__print_configuration__()
+
+
+class ConfigDispatcher:
+    def __init__(self, cl_config: ConfigProgram = None):
+        self.max_msg_len = 50000
+        self.alive_message = 24
+
+        self.telegram_enable = False
+        self.telegram_chat_id = '0'
+        self.telegram_token = ''
+        self.telegram_max_msg_len = 2000
+        self.telegram_rate_limit = 20
+
+        self.email_enable = True
+        self.email_smtp_server = ''
+        self.email_smtp_port = 587
+        self.email_sender = ''
+        self.email_sender_password = '***'
+        self.email_recipient = ''
+
+        if cl_config is not None:
+            self.__init_configuration_app__(cl_config)
+
+    def __mask_data__(self, data):
+        if len(data) > 2:
+            index = int(len(data) / 2)
+            partial = '*' * index
+            return f"{self.telegram_token[:index]}{partial}"
+        else:
+            return data
+
+    def __print_configuration__(self):
+        """
+        Print setup class
+        """
+
+        print(f"INFO    [Dispatcher setup] telegram={self.telegram_enable}")
+        if self.telegram_enable:
+            print(f"INFO    [Dispatcher setup] telegram-chat id={self.telegram_chat_id}")
+            # if len(self.telegram_token) > 2:
+            #     index = int(len(self.telegram_token) / 2)
+            #     partial = '*' * index
+            #     print(f"INFO    [Dispatcher setup] telegram-token={self.telegram_token[:index]}{partial}")
+            # else:
+            #     print(f"INFO    [Dispatcher setup] telegram-token={self.telegram_token}")
+            print(f"INFO    [Dispatcher setup] telegram-token={self.__mask_data__(self.telegram_token)}")
+
+            print(f"INFO    [Dispatcher setup] telegram-max message length={self.telegram_max_msg_len}")
+            print(f"INFO    [Dispatcher setup] telegram-rate limit minute={self.telegram_rate_limit}")
+            print(f"INFO    [Dispatcher setup] Notification-alive message every={self.alive_message} hour")
+
+        print(f"INFO    [Dispatcher setup] email={self.email_enable}")
+        if self.email_enable:
+            print(f"INFO    [Dispatcher setup] email-smtp server={self.email_smtp_server}")
+            print(f"INFO    [Dispatcher setup] email-port={self.email_smtp_port}")
+            print(f"INFO    [Dispatcher setup] email-sender={self.email_sender}")
+            # print(f"INFO    [Dispatcher setup] email-password={self.__mask_data__(self.email_sender_password)}")
+            print(f"INFO    [Dispatcher setup] email-password={self.email_sender_password}")
+            print(f"INFO    [Dispatcher setup] email-recipient={self.email_recipient}")
+
+    def __init_configuration_app__(self, cl_config: ConfigProgram):
+        """
+        Init configuration class reading .env file
+        """
+        # telegram section
+        self.telegram_enable = cl_config.telegram_enable()
+        self.telegram_chat_id = cl_config.telegram_chat_id()
+        self.telegram_token = cl_config.telegram_token()
+        self.telegram_max_msg_len = cl_config.telegram_max_msg_len()
+        self.telegram_rate_limit = cl_config.telegram_rate_limit_minute()
+
+        self.email_enable = cl_config.email_enable()
+        self.email_sender = cl_config.email_sender()
+        self.email_sender_password = cl_config.email_sender_password()
+
+        self.email_smtp_port = cl_config.email_smtp_port()
+        self.email_smtp_server = cl_config.email_smtp_server()
+        self.email_recipient = cl_config.email_recipient()
+
+        self.alive_message = cl_config.notification_alive_message_hours()
+
+        # email section
         self.__print_configuration__()